requirements.txt
discord_webhook==1.3.0
google_api_python_client==2.100.0
protobuf==4.24.3
schedule==1.2.0
